refactor: demote RichValueHelper to module-level functions in rich_value.py
Per issue #134's helper-consolidation plan (item: rich_value_helper -
"Demote... 202 lines of static methods... class wrapper buys nothing.
Drop the class, keep functions in a renamed file.").

RichValueHelper was already 100% @staticmethod/@classmethod with no
instance state, called via RichValueHelper.method(...) or cls.method(...)
everywhere. Converted process_rich_values, _is_rich_value, and
_is_legacy_rich_value to plain module-level functions (keeping the
underscore-prefix convention for the two internal-use ones) in a new
rich_value.py, and updated all 3 call sites (structured_model.py,
comparison_engine.py comment, confidence/accumulator.py).

Caught a real naming collision before it became a bug: structured_model.py's
from_json() has a boolean parameter also named process_rich_values. Once
imported directly, the module-level function would have been shadowed by
that parameter inside from_json() (Python scoping treats a name assigned
anywhere in a function as local throughout it), turning the call at the
unwrap site into `True(json_data)` - a TypeError at runtime that no static
check would have caught before actually exercising that code path. Imported
the function there under an alias (`unwrap_rich_values`) instead of renaming
the public `process_rich_values` parameter.

Testing: unlike every other item in this series, this module already had a
substantial dedicated test file (test_rich_value_pattern.py, 24 direct
references including to the "private" _is_rich_value/_is_legacy_rich_value
functions) plus a docstring mention in test_prediction_raw_roundtrip.py -
both updated to import and call the new module-level functions directly.
Full suite passes unchanged: 1541 passed, 2 pre-existing unrelated skips.

Updates 7 docs: the 3 usual suspects plus 4 more found only via the
exhaustive repo-wide search - a user-facing guide
(docs/Advanced/confidence-metrics.md) and two subdirectory READMEs
(models/confidence/README.md, models/bbox/README.md) that referenced this
class in prose, which the earlier, narrower doc searches in this series
hadn't surfaced.

Contributes to #134 (7 of 13 helpers; 6 remain, plus field_helper.py
deferred due to a conflicting open PR #183).
